requirements.txt
beautifulsoup4==4.13.3
groq==0.18.0
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.31.0
